requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
pytz>=2024.1
mysql-connector-python>=8.3.0 